refactor(app-header): split into mobile nav, desktop nav and actions
Extracts the header's mobile sheet, desktop navigation and right-side actions
into dedicated components with a shared nav-items module, keeping app-header
under the size limits. Refs #81.
