+++ resources/js/components/app-header/nav-items.ts
@@ -1,0 +1,26 @@
+import { type NavItem } from '@/types';
+import { BookOpen, Folder, LayoutGrid } from 'lucide-react';
+
+export const mainNavItems: NavItem[] = [
+    {
+        title: 'Dashboard',
+        url: '/dashboard',
+        icon: LayoutGrid,
+    },
+];
+
+export const rightNavItems: NavItem[] = [
+    {
+        title: 'Repository',
+        url: 'https://github.com/laravel/react-starter-kit',
+        icon: Folder,
+    },
+    {
+        title: 'Documentation',
+        url: 'https://laravel.com/docs/starter-kits',
+        icon: BookOpen,
+    },
+];
+
+export const activeItemStyles =
+    'text-neutral-900 dark:bg-neutral-800 dark:text-neutral-100';
